--- dialogue-core/src/test/java/com/palantir/dialogue/core/TestHostAndLimitedChannels.java
@@ -1,55 +1,0 @@
-/*
- * (c) Copyright 2021 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.dialogue.core;
-
-import com.google.common.collect.ImmutableList;
-import com.palantir.dialogue.DialogueImmutablesStyle;
-import java.util.stream.IntStream;
-import org.immutables.value.Value;
-
-public final class TestHostAndLimitedChannels {
-
-    private TestHostAndLimitedChannels() {}
-
-    static HostAndLimitedChannels createAndAssignHostIdx(ImmutableList<LimitedChannel> limitedChannels) {
-        ImmutableList<HostAndLimitedChannel> channels = IntStream.range(0, limitedChannels.size())
-                .mapToObj(HostIdx::of)
-                .map(hostIdx -> TestHostAndLimitedChannel.builder()
-                        .hostIdx(hostIdx)
-                        .limitedChannel(limitedChannels.get(hostIdx.index()))
-                        .build())
-                .collect(ImmutableList.toImmutableList());
-        return HostAndLimitedChannels.create(channels);
-    }
-
-    static HostAndLimitedChannel hostAndLimitedChannel(int hostIdx, LimitedChannel limitedChannel) {
-        return TestHostAndLimitedChannel.builder()
-                .hostIdx(HostIdx.of(hostIdx))
-                .limitedChannel(limitedChannel)
-                .build();
-    }
-
-    @Value.Immutable
-    @DialogueImmutablesStyle
-    interface TestHostAndLimitedChannel extends HostAndLimitedChannel {
-        class Builder extends ImmutableTestHostAndLimitedChannel.Builder {}
-
-        static Builder builder() {
-            return new Builder();
-        }
-    }
-}
